Attempt to prevent users from running script on an already merged repo
This commit introduces a fail safe to prevent users from getting
merge conflicts when merging the commits into their repos. Logic
is as follows:

1. Gets the commit message from the second hash found (which will
   be the first commit picked)
2. Strips it off any brackets (the only symbol that will cause a
   grep to fail)
3. Attempts to find it with git log; if it is found, the script
   will error and exit. If it isn't, the script will proceed.

There should not be a way for this to fail, unless the user
manually picked the commits (but they wouldn't be running this
script anyways lol).

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>